feat: Enhance RNCP functionality by integrating CalculatorTable
