+++ data/rncp_21.json
@@ -275,7 +275,15 @@
       1662,
       1880,
       1864,
-      1872
+      1872,
+      2561,
+      2562,
+      2563,
+      2564,
+      2338,
+      2339,
+      2340,
+      2341
     ]
   }
 }
